move updatePlayerType hook into its own patch
