+++ app/src/main/java/app/revanced/integrations/patches/PlayerTypeHookPatch.java
@@ -1,0 +1,32 @@
+package app.revanced.integrations.patches;
+
+import androidx.annotation.Nullable;
+
+import app.revanced.integrations.utils.LogHelper;
+import app.revanced.integrations.utils.PlayerType;
+
+/**
+ * Hook receiver class for 'player-type-hook' patch
+ *
+ * @usedBy app.revanced.patches.youtube.misc.playertype.patch.PlayerTypeHookPatch
+ * @smali Lapp/revanced/integrations/patches/PlayerTypeHookPatch;
+ */
+@SuppressWarnings("unused")
+public class PlayerTypeHookPatch {
+    /**
+     * Hook into YouTubePlayerOverlaysLayout.updatePlayerLayout() method
+     *
+     * @param type the new player type
+     * @smali YouTubePlayerOverlaysLayout_updatePlayerTypeHookEX(Ljava/lang/Object;)V
+     */
+    public static void YouTubePlayerOverlaysLayout_updatePlayerTypeHookEX(@Nullable Object type) {
+        if (type == null) return;
+
+        // update current player type
+        final PlayerType newType = PlayerType.safeParseFromString(type.toString());
+        if (newType != null) {
+            PlayerType.setCurrent(newType);
+            LogHelper.debug(PlayerTypeHookPatch.class, "YouTubePlayerOverlaysLayout player type was updated to " + newType);
+        }
+    }
+}
